+++ gettext.php
@@ -4,7 +4,7 @@
  *
  * @copyright (c) Mateo Tibaquirá
  * @link      http://www.blanktheme.org
- * @license   GNU/GPL - http://www.gnu.org/copyleft/gpl.html
+ * @license   MIT - http://www.opensource.org/licenses/mit-license.html
  * @abstract  BlankTheme gettext string to translate
  */
 
@@ -15,7 +15,6 @@ no__('no');
 no__('Master body template');
 no__('2 columns');
 no__('3 columns');
-no__('3 columns static');
 no__('grid');
 no__('full width');
 
@@ -24,7 +23,7 @@ no__('Master layout');
 no__("Enable 'header' block position?");
 
 no__('User menu to use');
-no__('bt_userlinks plugin');
+no__('blankmenu plugin');
 no__("'topnav' block position");
 
 no__('Bottom navigation to use');
@@ -34,5 +33,3 @@ no__("'bottomnav' block position");
 no__("Enable 'footer' block position?");
 
 no__('Enable font resize buttons');
-
-no__('Use the layout_X.css');